requirements.txt
streamlit==1.22.0
pickleshare==0.7.5
scikit-learn==1.2.2
Pillow==9.5.0
